[backend] refactor: 일기 수정/상세조회 로직 계층 분리 및 구조 개선 (Controller → Service → Repository → Mapper)

+++ backend/src/main/java/com/example/demo/team/TeamMapper.java
@@ -1,7 +1,5 @@
 package com.example.demo.team;
 
-import com.example.demo.request.CreateTeamRequest;
-import com.example.demo.request.FindTeamIdRequestModel;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
 
@@ -24,7 +22,8 @@ public interface TeamMapper {
     void deleteTeam(int teamId);
 
     List<TeamModel> searchTeam(String searchWord);
-    List<TeamModel> searchTeamByDiaryId(int diaryId);
+
+    List<TeamModel> searchTeamByDiaryId(long diaryId);
 
     // 팀 id 조회
     List<TeamModel> findTeamId(String teamName, int creatorId);
